fix cargo dev-loop
